refactor: refactor project code structure
